requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pyjwt[crypto]